Created useSubmitJob.ts, set up job to run similarly to compress hook
