+++ client/src/hooks/datafiles/mutations/useExtract.ts
@@ -3,64 +3,8 @@ import { useSelector, useDispatch, shallowEqual } from 'react-redux';
 import { getExtractParams } from 'utils/getExtractParams';
 import { apiClient } from 'utils/apiClient';
 import { fetchUtil } from 'utils/fetchUtil';
-import {
-  TTapisSystem,
-  TAppFileInput,
-  TTapisJob,
-  TJobArgSpecs,
-  TJobKeyValuePair,
-  TTapisFile,
-} from 'utils/types';
-
-export type TJobPostOperations = 'resubmitJob' | 'cancelJob' | 'submitJob';
-
-export type TParameterSetSubmit = {
-  appArgs?: TJobArgSpecs;
-  containerArgs?: TJobArgSpecs;
-  schedulerOptions?: TJobArgSpecs;
-  envVariables?: TJobKeyValuePair[];
-};
-
-export type TConfigurationValues = {
-  execSystemId?: string;
-  execSystemLogicalQueue?: string;
-  maxMinutes?: number;
-  nodeCount?: number;
-  coresPerNode?: number;
-  allocation?: string;
-  memoryMB?: number;
-};
-
-export type TOutputValues = {
-  name?: string;
-  archiveSystemId?: string;
-  archiveSystemDir?: string;
-};
-
-export interface TJobSubmit extends TConfigurationValues, TOutputValues {
-  archiveOnAppError?: boolean;
-  appId?: string;
-  fileInputs?: TAppFileInput[];
-  parameterSet?: TParameterSetSubmit;
-}
-
-export type TJobBody = {
-  operation?: TJobPostOperations;
-  uuid?: string;
-  job: TJobSubmit;
-  licenseType?: string;
-  isInteractive?: boolean;
-  execSystemId?: string;
-};
-
-interface IJobPostResponse extends TTapisJob {
-  execSys?: TTapisSystem;
-}
-
-type TJobPostResponse = {
-  response: IJobPostResponse;
-  status: number;
-};
+import { TTapisFile } from 'utils/types';
+import { TJobBody, TJobPostResponse } from './useSubmitJob';
 
 const getAppUtil = async function fetchAppDefinitionUtil(
   appId: string,
